+++ CHANGELOG.md
@@ -6,7 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
-## xxx
+## Unreleased
+
+### Added
+
+- Manage and apply templates to lists ([#255](https://github.com/garritfra/fling/pull/255))
 
 ## v0.8.5 (2024-07-16)